+++ src/test_vectors.rs
@@ -44,6 +44,7 @@ struct _Ed25519TestVector {
 
 struct _MoneroTestVector {
   mnemonic_words: &'static str,
+  mnemonic_passphrase: &'static str,
   derivation_path: DerivationPathData,
   expected_monero_address: &'static str,
 }
@@ -886,7 +887,8 @@ mod tests {
     let mut wallet = CryptoWallet::new();
 
     let test_vectors = vec![_MoneroTestVector {
-      mnemonic_words: "credit chat quit fine fashion poet drum budget number cart acoustic flower holiday patrol potato buffalo sorry latin sibling digital top dolphin zebra silly",
+      mnemonic_words: "permit universe parent weapon amused modify essay borrow tobacco budget walnut lunch consider gallery ride amazing frog forget treat market chapter velvet useless topple",
+      mnemonic_passphrase: "a",
       derivation_path: DerivationPathData {
         purpose: Zeroizing::new(44),
         purpose_hardened: Zeroizing::new(true),
@@ -905,11 +907,11 @@ mod tests {
 
         last_index: Zeroizing::new(0),
       },
-      expected_monero_address: "4A4G1vY7m7DjUoNcXMAAeWHaKWAPrjA1yWNvxR7EvaddHdzb1tx5HQBZoKJA2QfgAbAqPtY9T7oqWRv8MjnrCxeKMoYvrJC",
+      expected_monero_address: "4BBKEeg8iH3JtyQKfdh5KRYNe2WK4aDMBeMwPvkjrm45BQ8bVHurmLyadDx3EiM6AjNH7JJx5TMNrjC4JLZEhszc5f3G8Yg",
     }];
 
     for vector in test_vectors {
-      let seed_raw = match generate_seed_from_mnemonic(vector.mnemonic_words, None) {
+      let seed_raw = match generate_seed_from_mnemonic(vector.mnemonic_words, Some(vector.mnemonic_passphrase)) {
         Ok(seed) => seed,
         Err(_) => {
           panic!("Can not generate seed from mnemonic");
@@ -923,6 +925,7 @@ mod tests {
         }
       };
 
+      assert_eq!("ca23c96aa8552adf211ebbe1d23f78670b5ffb541b5276e26db1b2c7943ae7b354b013805b1836390187a1a3d5915f109a0f7a67c4e4603bb451a5df12bb5931", &seed_hex);
       wallet.seed_secret.seed = Zeroizing::new(seed_hex);
 
       let _ = keys::generate_ed25519_master_keys(&mut wallet);
@@ -944,12 +947,11 @@ mod tests {
           .child_private_key_bytes,
       );
 
-      use crate::dev;
-      let spend_priv = dev::monero_sc_reduce32(&spend_raw).to_bytes();
-      let view_priv = dev::monero_secret_view_key(&spend_priv);
-      let spend_pub = dev::monero_pubkey(&spend_priv);
-      let view_pub = dev::monero_pubkey(&view_priv);
-      let address = dev::generate_monero_address(&spend_pub, &view_pub);
+      let spend_priv = crate::dev::monero_sc_reduce32(&spend_raw).to_bytes();
+      let view_priv = crate::dev::monero_secret_view_key(&spend_priv);
+      let spend_pub = crate::dev::monero_pubkey(&spend_priv);
+      let view_pub = crate::dev::monero_pubkey(&view_priv);
+      let address = crate::dev::generate_monero_address(&spend_pub, &view_pub);
 
       assert_eq!(address, vector.expected_monero_address);
     }
@@ -1104,19 +1106,19 @@ fn check_wallet_save_open_function() {
 // Monero Debugging
 // monero-seed --slip0010 --passphrase 'a' -- 'permit universe parent weapon amused modify essay borrow tobacco budget walnut lunch consider gallery ride amazing frog forget treat market chapter velvet useless topple'
 // # OR: python3 -m monero_poc.seed --slip0010 --passphrase 'a' -- 'permit universe parent weapon amused modify essay borrow tobacco budget walnut lunch consider gallery ride amazing frog forget treat market chapter velvet useless topple'
-//
+// 
 // Seed bip39 words: permit universe parent weapon amused modify essay borrow tobacco budget walnut lunch consider gallery ride amazing frog forget treat market chapter velvet useless topple
 // Seed bip32 b58:   ca23c96aa8552adf211ebbe1d23f78670b5ffb541b5276e26db1b2c7943ae7b354b013805b1836390187a1a3d5915f109a0f7a67c4e4603bb451a5df12bb5931
-//
+// 
 // Seed Monero:      0b86cf0e71204ca3cdc389daf0bb1cf654ac7d54edfed68a9faf921ba140a708
 // Seed Monero wrds: maul loudly nearby buffet hacksaw zones kernels edgy baffles match extra eclipse uphill arena hounded wobbly actress muppet pebbles onward rift tether scrub snake rift
-//
+// 
 // Private spend key: 0b86cf0e71204ca3cdc389daf0bb1cf654ac7d54edfed68a9faf921ba140a708
 // Private view key:  3b1213f644062fbfb15c4fa35e656659e4105ac728b791ce5ad10af98bc6d200
-//
+// 
 // Public spend key:  fc25f0a1a1fb4e6afe5ffa15cd06fcbb913eb55605d09adf5de735163d4f4a3e
 // Public view key:   2ba093948b429ec90729ecd0f705b87f36154612756433fc3dc7d8dbbf646529
-//
+// 
 // Mainnet Address:   4BBKEeg8iH3JtyQKfdh5KRYNe2WK4aDMBeMwPvkjrm45BQ8bVHurmLyadDx3EiM6AjNH7JJx5TMNrjC4JLZEhszc5f3G8Yg
 // Testnet Address:   A2iriuLPze9JtyQKfdh5KRYNe2WK4aDMBeMwPvkjrm45BQ8bVHurmLyadDx3EiM6AjNH7JJx5TMNrjC4JLZEhszc5dqa8Po
 // Stagenet Address:  5BPMKVb6Mt9JtyQKfdh5KRYNe2WK4aDMBeMwPvkjrm45BQ8bVHurmLyadDx3EiM6AjNH7JJx5TMNrjC4JLZEhszc5eeaUED