+++ doc/changelog.md
@@ -1,5 +1,8 @@
 # version 1.5.1
 - new coin: Monero (XMR)
+- new coin: Zilliqa (ZIL)
+    - Legacy address zil1... supported
+    - New V2 address 0x... supported
 
 ---